feat: updated the code to the latest pattern, and included a STETH/DAI and STETH/USDC price feeds
